Main code
This is the actual code for generating the username and password. The reason for all the comments is mostly for the "diary keeping" of learning Python for me. At some point I will clean up this portfolio and have these learning projects in their own groups, away from the larger projects that are coming out in the near future.